+++ src/apis/models/payment/model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import JSON, Column, DateTime, ForeignKey, Index, Integer, String, UniqueConstraint, text
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Index, Integer, String, UniqueConstraint, text
 
 from services.database.postgres.connection import Base
 
@@ -14,8 +14,8 @@ class Payment(Base):
         Index("ix_payments_order_id", "order_id"),
     )
 
-    id = Column(Integer, primary_key=True, index=True)
-    order_id = Column(Integer, ForeignKey("orders.id"), nullable=False)
+    id = Column(BigInteger, primary_key=True, index=True)
+    order_id = Column(BigInteger, ForeignKey("orders.id"), nullable=False)
     provider = Column(String, nullable=False)
     provider_order_id = Column(String, nullable=False)
     provider_payment_id = Column(String, nullable=True)
