+++ src/apis/models/session/model.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, DateTime, ForeignKey, Index, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Index, String
 
 from services.database.postgres.connection import Base
 
@@ -9,8 +9,8 @@ class Session(Base):
     __tablename__ = "sessions"
     __table_args__ = (Index("ix_sessions_user_id_revoked_at", "user_id", "revoked_at"),)
 
-    id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
+    id = Column(BigInteger, primary_key=True, index=True)
+    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
     refresh_token_hash = Column(String, unique=True, nullable=False, index=True)
     device_info = Column(String, nullable=True)
     ip_address = Column(String, nullable=True)
